initial implementation of {url} placeholder
In addition to {filepath}, {url} allows upload commands
to get the current metadb/hashed artwork URL (if present),
this way a script can verify whether URLs are still valid
and if not reupload them (and return a valid URL to update,
or nothing valid as a NOOP). This placeholder will result
in the command being executed every single time.